requirements.txt
# VisualLink: Missing Person Reunification Hub
# Python Dependencies

# Google ADK (Agent Development Kit)
google-adk>=0.1.0
google-generativeai>=0.8.0

# Firebase / Firestore
firebase-admin>=6.5.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0

# Google Workspace APIs
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Image Processing
Pillow>=10.4.0
opencv-python-headless>=4.10.0
numpy>=1.26.0

# Video Processing
imageio>=2.34.0
imageio-ffmpeg>=0.5.1

# Frontend
streamlit>=1.38.0
streamlit-option-menu>=0.3.13
streamlit-folium>=0.22.0
folium>=0.17.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.8.0
requests>=2.32.0
uuid>=1.30
python-dateutil>=2.9.0
tenacity>=9.0.0

# Async support
asyncio-mqtt>=0.16.2
aiofiles>=24.1.0

# Logging & Monitoring
structlog>=24.4.0
rich>=13.8.0
